nit: Moves EdgeMetaData to thor/path_algorithm.h
