+++ valhalla/thor/bidirectional_astar.h
@@ -35,8 +35,6 @@ struct CandidateConnection {
   }
 };
 
-struct EdgeMetadata;
-
 /**
  * Bidirectional A* algorithm. Method for finding least-cost path.
  */
